adjusting swagger

+++ src/controllers/profiles_controller.py
@@ -19,11 +19,24 @@ from schemas.ProfileImagesSchema import profile_image_schema, profile_images_sch
 
 profile = Blueprint("profile", __name__, url_prefix="/profile")
 
-tables = [Equipment, EquipmentOrder, Profile, ProfileImages, User]
+tables = ["equipment", "equipment_order", "profiles", "profile_images", "users"]
 schemas = [equipments_schema, equipment_orders_schema, profiles_schema, profile_images_schema, users_schema ]
 
-@profile.route("/dump/all", methods=["GET"])
-def profile_dump():
+
+
+
+@profile.route("dump/all/<int:id>", methods=["GET"])
+@jwt_required
+@verify_user
+def profile_dump(user, id):
+
+    profile = db.session.query(Profile).filter(Profile.admin == True).filter_by(profileid = id, user_id=user.id).first()
+
+    
+
+
+    if not profile:
+        return abort(400, description="Unauthorised to complete this action")
     i=0
     try:
         os.remove("backup/backup.json")
@@ -31,10 +44,14 @@ def profile_dump():
     except:
         print("file does not exist")
     for table in tables:
-
-
-        query = db.session.query(table)
+        
+        
+        query = db.engine.execute(f'SELECT * FROM {table}')
         data = ((schemas[i]).dump(query))
+
+
+        print(data)
+
         data = json.dumps(data)
         i+=1
     
@@ -48,9 +65,6 @@ def profile_dump():
     return "Data backed up"
 
 
-
-
-
 @profile.route("/all", methods=["GET"])
 def profile_index():
     query = db.session.query(Profile)
@@ -60,13 +74,13 @@ def profile_index():
 
 @profile.route("/active", methods=["GET"])
 def profile_index_active():
-    # query = session.query(Profile)
+ 
     query = db.session.query(Profile).filter(Profile.account_active).order_by(Profile.fname)
     return jsonify(profiles_schema.dump(query))
 
 @profile.route("/onhire", methods=["GET"])
 def profile_index_profile_equipment_on_hire():
-    # query = session.query(Profile)
+
     query = db.session.query(Profile, Equipment, EquipmentOrder).outerjoin(EquipmentOrder, Profile.profileid == EquipmentOrder.hirer_id).outerjoin(Equipment, EquipmentOrder.hirer_id == Equipment.owner_id).order_by(Profile.username).all()
     things = []
     for result in query:
@@ -76,13 +90,12 @@ def profile_index_profile_equipment_on_hire():
 
 @profile.route("/equipment", methods=["GET"])
 def profile_index_profile_equipment_rent():
-    # query = session.query(Profile)
+    
     query = db.session.query(Profile, Equipment).join(Equipment, Profile.profileid == Equipment.owner_id ).order_by(Profile.username).all()
     things = []
     for result in query:
         things.append(f"Name: {result[0].username} Rentpw: {result[1].rentpw}")
-    print(things)
-    return jsonify((things))
+    return jsonify(things)
 
 
 
@@ -107,6 +120,7 @@ def profile_create(user=None):
         new_profile.fname = profile_fields["fname"]
         new_profile.lname = profile_fields["lname"]
         new_profile.account_active=profile_fields["account_active"]
+        new_profile.admin=profile_fields["admin"]
         
         user.profile.append(new_profile)
         
@@ -125,32 +139,32 @@ def profile_show(username):
     profile = Profile.query.filter_by(username = username).first()
     return jsonify(profile_schema.dump(profile))
 
-@profile.route("/<string:username>", methods=["PUT", "PATCH"])
+@profile.route("/<int:id>", methods=["PUT", "PATCH"])
 @jwt_required
 @verify_user
-def profile_update(username, user=None):
+def profile_update(user, id):
 
 
-    profile = Profile.query.filter_by(username = username, user_id=user.id)
-
+    profile = Profile.query.filter_by(profileid = id, user_id=user.id)
+    
     profile_fields = profile_schema.load(request.json)
 
-    if profile.count() != 1:
+    if not profile:
         return abort(401, description="Unauthorised to update this user")
+    
+    
     profile.update(profile_fields)
-
-
     db.session.commit()
-
+    )
     return jsonify(profile_schema.dump(profile[0]))
 
-@profile.route("/<string:username>", methods=["DELETE"])
+@profile.route("/<int:id>", methods=["DELETE"])
 @jwt_required
 @verify_user
-def profile_delete(username, user=None):
+def profile_delete(user, id):
 
 
-    profile = Profile.query.filter_by(username = username, user_id=user.id).first()
+    profile = Profile.query.filter_by(profileid = id, user_id=user.id).first()
 
 
     if not profile:
